+++ CHANGELOG.md
@@ -9,6 +9,8 @@ You can also check [on GitHub](https://github.com/nextcloud/news/releases), the 
 ### Changed
 
 ### Fixed
+- style fixes for `<blockquote>`, `<pre>` and `<code>` in articles
+- keep action bar position when scrolling article
 
 # Releases
 ## [26.0.1] - 2025-06-04